# openapi/lis-internal-v1.yaml
openapi: 3.1.0
info:
  title: LIS Core API
  summary: Starter modular monolith for a laboratory information system
  description: Internal API with working persistence for auth, master data, orders,
    specimen lifecycle, task orchestration, observations, reports, audit, provenance,
    device registry, HL7 v2 integrations, device gateway, autoverification, ASTM-style
    drivers, and a read/search FHIR facade.
  version: 0.7.0
paths:
...
  /fhir/R4/Patient:
    get:
      tags:
      - fhir
      summary: Search Patients
      operationId: search_patients_fhir_R4_Patient_get
      security:
      - HTTPBearer: []
      parameters:
      - name: identifier
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Identifier
      - name: family
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Family
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Patients Fhir R4 Patient Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/Patient/{patient_id}:
    get:
      tags:
      - fhir
      summary: Read Patient
      operationId: read_patient_fhir_R4_Patient__patient_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Patient Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Patient Fhir R4 Patient  Patient Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/ServiceRequest:
    get:
      tags:
      - fhir
      summary: Search Service Requests
      operationId: search_service_requests_fhir_R4_ServiceRequest_get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Patient
      - name: identifier
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Identifier
      - name: requisition
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Requisition
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Service Requests Fhir R4 Servicerequest Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /fhir/R4/Specimen:
    get:
      tags:
      - fhir
      summary: Search Specimens
      operationId: search_specimens_fhir_R4_Specimen_get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Patient
      - name: identifier
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Identifier
      - name: accession
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Accession
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Specimens Fhir R4 Specimen Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /fhir/R4/Task:
    get:
      tags:
      - fhir
      summary: Search Tasks
      operationId: search_tasks_fhir_R4_Task_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: focus
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Focus
      - name: code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
      - name: patient
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Patient
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Tasks Fhir R4 Task Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /fhir/R4/Observation:
    get:
      tags:
      - fhir
      summary: Search Observations
      operationId: search_observations_fhir_R4_Observation_get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Patient
      - name: specimen
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Specimen
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
      - name: based-on
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Based-On
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Observations Fhir R4 Observation Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /fhir/R4/DiagnosticReport:
    get:
      tags:
      - fhir
      summary: Search Diagnostic Reports
      operationId: search_diagnostic_reports_fhir_R4_DiagnosticReport_get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Patient
      - name: based-on
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Based-On
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Diagnostic Reports Fhir R4 Diagnosticreport
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /fhir/R4/AuditEvent:
    get:
      tags:
      - fhir
      summary: Search Audit Events
      operationId: search_audit_events_fhir_R4_AuditEvent_get
      security:
      - HTTPBearer: []
      parameters:
      - name: entity
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Audit Events Fhir R4 Auditevent Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /fhir/R4/Provenance:
    get:
      tags:
      - fhir
      summary: Search Provenance
      operationId: search_provenance_fhir_R4_Provenance_get
      security:
      - HTTPBearer: []
      parameters:
      - name: target
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Target
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Provenance Fhir R4 Provenance Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/auth/bootstrap-admin:
    post:
      tags:
      - auth
      summary: Bootstrap Admin
      operationId: bootstrap_admin_api_v1_auth_bootstrap_admin_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapAdminRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/patients:
    post:
      tags:
      - patients
      summary: Create Patient
      operationId: create_patient_api_v1_patients_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - patients
      summary: List Patients
      operationId: list_patients_api_v1_patients_get
      security:
      - HTTPBearer: []
      parameters:
      - name: mrn
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Mrn
      - name: family_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Family Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/PatientSummary'
                title: Response List Patients Api V1 Patients Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/test-catalog:
    post:
      tags:
      - catalog
      summary: Create Test Catalog Item
      operationId: create_test_catalog_item_api_v1_test_catalog_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCatalogCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCatalogSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - catalog
      summary: List Catalog
      operationId: list_catalog_api_v1_test_catalog_get
      security:
      - HTTPBearer: []
      parameters:
      - name: local_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Local Code
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/TestCatalogSummary'
                title: Response List Catalog Api V1 Test Catalog Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/orders:
    post:
      tags:
      - orders
      summary: Create Order
      operationId: create_order_api_v1_orders_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - orders
      summary: List Orders
      operationId: list_orders_api_v1_orders_get
      security:
      - HTTPBearer: []
      parameters:
      - name: requisition_no
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Requisition No
      - name: patient_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Patient Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderSummary'
                title: Response List Orders Api V1 Orders Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/{order_id}:
    get:
      tags:
      - orders
      summary: Get Order
      operationId: get_order_api_v1_orders__order_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/specimens/accession:
    post:
      tags:
      - specimens
      summary: Accession Specimen
      operationId: accession_specimen_api_v1_specimens_accession_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessionSpecimenRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecimenSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/specimens/{specimen_id}/collect:
    post:
      tags:
      - specimens
      summary: Collect Specimen
      operationId: collect_specimen_api_v1_specimens__specimen_id__collect_post
      security:
      - HTTPBearer: []
      parameters:
      - name: specimen_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Specimen Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectSpecimenRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecimenSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/specimens/{specimen_id}/receive:
    post:
      tags:
      - specimens
      summary: Receive Specimen
      operationId: receive_specimen_api_v1_specimens__specimen_id__receive_post
      security:
      - HTTPBearer: []
      parameters:
      - name: specimen_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Specimen Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiveSpecimenRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecimenSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/specimens/{specimen_id}/accept:
    post:
      tags:
      - specimens
      summary: Accept Specimen
      operationId: accept_specimen_api_v1_specimens__specimen_id__accept_post
      security:
      - HTTPBearer: []
      parameters:
      - name: specimen_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Specimen Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecimenSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/tasks/{task_id}/start:
    post:
      tags:
      - tasks
      summary: Start Task
      operationId: start_task_api_v1_tasks__task_id__start_post
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/TaskActionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/tasks/{task_id}/complete:
    post:
      tags:
      - tasks
      summary: Complete Task
      operationId: complete_task_api_v1_tasks__task_id__complete_post
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/TaskActionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BootstrapAdminRequest:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          minLength: 8
          title: Password
        display_name:
          type: string
          title: Display Name
      additionalProperties: false
      type: object
      required:
      - username
      - password
      - display_name
      title: BootstrapAdminRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReceiveSpecimenRequest:
      properties:
        received_at:
          type: string
          format: date-time
          title: Received At
        location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
      additionalProperties: false
      type: object
      required:
      - received_at
      title: ReceiveSpecimenRequest
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        user:
          $ref: '#/components/schemas/UserSummary'
      additionalProperties: false
      type: object
      required:
      - access_token
      - user
      title: TokenResponse
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer